packages/assistant/pyproject.toml
[project]
name = "voice-assistant"
version = "0.17.30"
description = "Developer-centric AI assistant for managing multiple software projects with TUI and optional voice interface"
authors = [{name = "xSwarm", email = "support@xswarm.io"}]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "numpy>=2.1.0",
    "scipy>=1.11.0",
    "sounddevice>=0.5.0",
    "sentencepiece>=0.2.0",
    "textual>=0.47.0",
    "rich>=13.7.0",
    "vosk>=0.3.44",
    "httpx>=0.26.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
    "pyperclip>=1.8.2",
    "huggingface-hub>=0.20.0",
    "hf-transfer>=0.1.0",  # Fast model downloads (50-100x speedup)
    "backoff>=1.11.0",  # Retry logic with exponential backoff for robust downloads
    "anthropic>=0.18.0",  # Claude API for conversation
    "openai>=1.12.0",  # GPT API for fallback
    "silero-vad>=4.0.0",  # Voice Activity Detection
    "twilio>=8.0.0",  # Phone call integration
    "sendgrid>=6.11.0",  # Email integration
    "toml>=0.10.2",  # Config file parsing
    "libsql-experimental>=0.0.55",  # LibSQL with vector search for semantic memory
    "sentence-transformers>=2.2.0",  # Local CPU embeddings for semantic search (no API key needed)
    "moshi_mlx @ git+https://github.com/kyutai-labs/moshi.git#subdirectory=moshi_mlx",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-textual-snapshot>=1.0.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
    "pytest-bdd>=7.0.0",
]
nvidia = [
    "pynvml>=11.5.0",  # NVIDIA GPU management library
]
amd = [
    "amdsmi>=0.1.0",  # AMD GPU management library (experimental)
]

[project.scripts]
xswarm = "assistant.main:main"
assistant = "assistant.main:main"
voice-assistant = "assistant.main:main"
xswarm-download-models = "assistant.voice.download_models:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["assistant*"]
exclude = ["tmp*", "tests*"]

[tool.setuptools.package-data]
assistant = ["*.tcss", "*.css", "personas/**/*.yaml", "personas/**/*.md", "personas/**/*.txt"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
